requirements.txt
flask
openai-whisper
librosa
numpy
yt-dlp
